requirements.txt
tensorflow>=2.12.0
numpy>=1.23.0
pillow>=9.5.0
matplotlib>=3.7.0
scikit-learn>=1.2.0
tk
